Fix CI (#3)
* Fix CI

* Fix flakey parallel execution test

The original test was expecting the timing to be less than a fixed
value. However, on CI it took longer. This change adds a 
`test_execute_parallel_timing` test that writes a time to the context
for each running operation surround by sleeps and compares their 
timestamps to ensure they actually ran in parallel.

* Appease ruff

* Fix incompatibility with Tuple in Python 3.8